logging completed

+++ pytranscoder/ffmpeg.py
@@ -88,7 +88,8 @@ class FFmpeg(Processor):
 
         if proc.returncode == 0:
             # if we got here then everything went fine, so remove the transaction log
-            os.remove(str(self.log_path))
+            if os.path.exists(self.log_path):
+                os.remove(str(self.log_path))
             self.log_path = None
 
     def run(self, params, event_callback) -> Optional[int]:
